# apps/api/openapi.yaml
openapi: 3.0.1
info:
  title: In Midst My Life API
  description: |
    Dynamic interactive CV/résumé system with mask-based identity filtering and narrative generation.

    ## Key Features
    - **Profile Management**: CRUD operations for user profiles with temporal epochs and career stages
    - **Mask System**: Context-specific identity masks with cognitive/expressive/operational ontologies
    - **CV Components**: Experiences, educations, skills, projects, credentials, and attestations
    - **Narrative Generation**: AI-powered résumé narrative blocks with weighting and LLM ranking
    - **Export Formats**: JSON-LD (semantic web), PDF (résumé), Verifiable Credentials (W3C standard)
    - **GraphQL API**: Unified query interface with subscriptions support
    - **Authentication**: JWT bearer tokens with role-based access control
    - **Caching**: Redis-backed (with in-memory fallback) for taxonomy endpoints

    ## Authentication

    Most endpoints require JWT bearer token authentication. Include in request:
    ```
    Authorization: Bearer <token>
    ```

    Tokens include permission claims for fine-grained access control.

  version: 0.2.0
  contact:
    name: API Support
    url: https://github.com/anthropics/in-midst-my-life
  license:
    name: MIT
servers:
  - url: http://localhost:3001
    description: Local development server
  - url: https://api.inmidstmylife.com
    description: Production server
...
paths:
...
  /profiles/{id}/narratives:
    get:
      tags: [Profiles]
      summary: Get narrative history
      operationId: getNarratives
      security:
        - bearerAuth: [READ_NARRATIVE]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, approved, rejected]
      responses:
        '200':
          description: Narratives retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/NarrativeSnapshot'
...
components:
  schemas:
...
    NarrativeBlock:
      type: object
      properties:
        title:
          type: string
        body:
          type: string
        tags:
          type: array
          items:
            type: string
        weight:
          type: number
        templateId:
          type: string
    NarrativeSnapshot:
      type: object
      properties:
        id:
          type: string
        blocks:
          type: array
          items:
            $ref: '#/components/schemas/NarrativeBlock'
        status:
          type: string
          enum: [draft, approved, rejected]
        createdAt:
          type: string
          format: date-time
        approvedAt:
          type: string
          format: date-time
        approvedBy:
          type: string
        metadata:
          type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token with permission claims

    agentAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: AI agent token for autonomous operations
security:
  - bearerAuth: []